remove mavenLocal()
We shouldn't need this.  If people need to do local hacking, they can
add it in, but it shouldn't be in the `master` build.gradle, because the
master shouldn't depend on files in the local maven repo.

I've uploaded the latest gds_metrics_dropwizard to alphagov bintray,
which removes the last reason we needed this.